fix(desktop): split parked budgets presentation model
