+++ apps/desktop/src/components/settings/panes/OrganizationBudgetsPane.tsx
@@ -9,51 +9,21 @@ import { SettingsCardRow } from "@/components/settings/shared/SettingsCardRow";
 import { SettingsPageHeader } from "@/components/settings/shared/SettingsPageHeader";
 import { useOrganizationMembers } from "@/hooks/access/cloud/organizations/use-organization-members";
 import { useActiveOrganization } from "@/hooks/organizations/facade/use-active-organization";
-import type { OrganizationMemberRecord } from "@/lib/domain/organizations/organization-records";
-
-const COMPUTE_BUDGET_PCUS = 500;
-const TOTAL_COMPUTE_PCUS = 360;
-const AVAILABLE_COMPUTE_PCUS = 118;
-const USED_COMPUTE_PCUS = TOTAL_COMPUTE_PCUS - AVAILABLE_COMPUTE_PCUS;
-const LLM_BUDGET_CREDITS = 12000;
-const TOTAL_LLM_CREDITS = 12000;
-const AVAILABLE_LLM_CREDITS = 4600;
-const USED_LLM_CREDITS = TOTAL_LLM_CREDITS - AVAILABLE_LLM_CREDITS;
-const USAGE_POINTS: UsagePoint[] = [
-  { label: "Jun 18", compute: 12, llm: 390 },
-  { label: "Jun 19", compute: 28, llm: 840 },
-  { label: "Jun 20", compute: 23, llm: 760 },
-  { label: "Jun 21", compute: 41, llm: 1260 },
-  { label: "Jun 22", compute: 58, llm: 1710 },
-  { label: "Jun 23", compute: 47, llm: 1480 },
-  { label: "Jun 24", compute: 33, llm: 960 },
-];
-const USAGE_BY_SOURCE = [
-  {
-    label: "LLM and model usage",
-    description: "Model calls, gateway usage, and inference-backed tools.",
-    value: "7,400 LLM credits",
-    percent: Math.round((USED_LLM_CREDITS / LLM_BUDGET_CREDITS) * 100),
-  },
-  {
-    label: "Compute runtime",
-    description: "Hosted environments, local runtime bridges, and execution time.",
-    value: "188 PCUs",
-    percent: Math.round((188 / COMPUTE_BUDGET_PCUS) * 100),
-  },
-  {
-    label: "Agent sessions",
-    description: "Workspace orchestration and background session services.",
-    value: "54 PCUs",
-    percent: Math.round((54 / COMPUTE_BUDGET_PCUS) * 100),
-  },
-];
-const FALLBACK_PEOPLE = [
-  { name: "Pablo", email: "pablo@pablohansen.com" },
-  { name: "Alex", email: "alex@example.com" },
-  { name: "Maya", email: "maya@example.com" },
-  { name: "Jordan", email: "jordan@example.com" },
-];
+import {
+  AVAILABLE_COMPUTE_PCUS,
+  AVAILABLE_LLM_CREDITS,
+  COMPUTE_BUDGET_PCUS,
+  LLM_BUDGET_CREDITS,
+  TOTAL_COMPUTE_PCUS,
+  TOTAL_LLM_CREDITS,
+  USED_COMPUTE_PCUS,
+  USED_LLM_CREDITS,
+  USAGE_BY_SOURCE,
+  USAGE_POINTS,
+  buildBudgetPeople,
+  type BudgetPerson,
+  type UsagePoint,
+} from "@/lib/domain/settings/organization-budgets-presentation";
 
 export function OrganizationBudgetsPane() {
   const { activeOrganization, activeOrganizationId, organizationsQuery } = useActiveOrganization();
@@ -313,23 +283,6 @@ function OrganizationMemberLlmBudgets({ people }: { people: BudgetPerson[] }) {
   );
 }
 
-interface BudgetPerson {
-  name: string;
-  email: string;
-  usedPcus: number;
-  usedLlmCredits: number;
-  monthlyLlmBudgetCredits: number;
-  alertThresholdPercent: number;
-  computePercent: number;
-  llmBudgetPercent: number;
-}
-
-interface UsagePoint {
-  label: string;
-  compute: number;
-  llm: number;
-}
-
 function BudgetBalanceCard({
   label,
   available,
@@ -362,33 +315,6 @@ function BudgetBalanceCard({
       </div>
     </div>
   );
-}
-
-function buildBudgetPeople(members: OrganizationMemberRecord[]): BudgetPerson[] {
-  const source = members.length > 0
-    ? members.map((member) => ({
-        name: member.displayName || member.email,
-        email: member.email,
-      }))
-    : FALLBACK_PEOPLE;
-
-  return source.slice(0, 5).map((person, index) => {
-    const usedPcus = [72, 45, 31, 22, 12][index] ?? 8;
-    const usedLlmCredits = [2800, 1900, 1320, 880, 500][index] ?? 250;
-    const monthlyLlmBudgetCredits = [5000, 3000, 2500, 1800, 1000][index] ?? 1000;
-    return {
-      ...person,
-      usedPcus,
-      usedLlmCredits,
-      monthlyLlmBudgetCredits,
-      alertThresholdPercent: [80, 80, 75, 75, 50][index] ?? 80,
-      computePercent: Math.round((usedPcus / COMPUTE_BUDGET_PCUS) * 100),
-      llmBudgetPercent: Math.min(
-        100,
-        Math.round((usedLlmCredits / monthlyLlmBudgetCredits) * 100),
-      ),
-    };
-  });
 }
 
 function UsageAreaChart({ points }: { points: UsagePoint[] }) {
